Error handling / display in the Dev UI and CLI. (#315)
Axios will throw an Error instead of returning non-200s. This code returns those errors
as responses instead, so that the typical status code checking can be done further
downstream.

Rename `InternalError` to `GenkitToolsError`, and add the `cause` field. Will give a
nice chain of errors when applicable.

This PR removes a bunch of logging at the catch block, in favor of logging further
downstream. A few tests with the CLI seem reasonable: