[Go] Action type (#326)
Add a generic type for actions.

This PR also adds some support functions:

We will want log messages to print some identifier for the request
they are part of, when this code is run in a server. Putting a logger
in a context helps with that.

We will be putting a few other things in contexts, so provide a common
utility, contextKey[T], to make that easier. We expect all the
stored values to be pointers, so having fromContext return the zero
value for a missing key simplifies coding without losing information.