[Go] add action registry (#390)
Add a global registry for actions.

This requires a common type for all generic types Action[I, O]. We
define the `action` interface for that purpose. We give it a runJSON
method, which will be useful when actions are called as part of a
server.

There is no Plugin type. A plugin is any package that calls
genkit.RegisterAction in an init function.

That means that plugins don't have names available to the program, but
plugin names don't seem to be useful.

We also don't provide lazy initialization of plugins. It isn't clear
that that is useful in Go.

(Also, remove paths-ignore from the required workflow checks. A design
flaw/bug in GitHub Actions makes these checks block forever if they are
filtered out, preventing submission.)