[Go] export traces to TraceStores (#492)
Following the js implementation, support two different environments,
dev and prod. The dev TraceStore is always set to use the local
filesystem. Calling RegisterTraceStore sets the prod TraceStore.

The Init function builds the OpenTelemetry Tracer from the TraceStores,
and returns a shutdown function, which is needed to flush batched
traces. We could build the tracer on first use if we didn't need
to return the shutdown function.

Requiring an Init function is unfortunate, but we would probably
have needed one eventually to support configuration.

For #354.