[Go] changes to TraceStore.List (#504)
These changes are taken from the latest js version of the corresponding
code.

FileTraceStore.List sorts by modified time. Following the js,
it sorts the most recently written files first.

Support a continuation token, to resume listing from where it left
off. Following the javascript, this is a string so that implementations
can use it to encode whatever they need.

For #354.