Redact the genkit:input and genkit:output attributes from spans (#693)
* Redact the genkit:input and genkit:output attributes from spans
in the GCP plugin.

These attributes may contain PII information. Both, the input and
the output are still logged to Cloud Logging, where the user has
sophisticated ACLs that facilitate controlled access to this data.

Add the /http/status_code attribute to all error spans and set its
value to 599. This is a workaround that will mark error spans with
a red exclamation mark in Cloud Trace to better indicate an error
condition.