[Go] implement the Run function (#751)
Run executes a step within a flow. It creates a new trace span,
and also caches the result so that if the flow is resumed, the
argument function won't be called again.

Each call to Run may pass a function that returns a different
type, and those types need not be related to the `I` and `O` type
parameters of the flow. So we can't use a type parameter for the cache.
If we used `any`, then unmarshaling a saved cache as JSON would result in
types like `[]any` and `map[string]any`. Using `json.RawMessage` as
the cache value ensures that values are unmarshaled into the right
types, which we know inside the body of `Run`.

For #354.