[Go] use non-global registry for testing (#786)
Make a `registry` type, of which the global registry is just one
instance. There can be other registries in play. This enables
tests to each have their own registry, avoiding interference.

A registry now holds a tracingState. The global tracingState is just
the global registry's tracingState.

Reorganize tracingState to support this, and also, somewhat unrelatedly,
simplify it. It turns out that a new Tracer doesn't need to be created
each time a SpanProcessor is added, and that TracerProvider is already
goroutine-safe. So remove the tracingState mutex, and create the
TracerProvider when a new tracingState is created.

For #354.