[Go] add language-independent tests (#818)
Add tests that read JSON files that describe flows and expected results.

The goal is to run these tests for each genkit language, to make
sure all implementations behave the same.

There is currently a single test, whose expected trace came from
writing the flow in ts and calling `genkit flow:run`. This test
found a couple of bugs and discrepancies in the Go code, which are
also fixed here.

For #354.