[Go] begin getting sample to run with genkit tool (#27)
Add the first Go sample, which implements the "basic" flow
from samples/flow-sample1.

Attempting to run this with `genkit start -a ...` flushed out a few
bugs. This PR fixes enough so that the flow is listed in the left
nav of the UI.

The next step requires adding JSON schemas to the actions, which is a
large piece, so this is a good place to stop.