[ Go ] define Embedder and Retriever interfaces (#35)
The intent is that the actual implementations of embedders and
retrievers will implement these interfaces. They will be wrapped in
actions for invocation from the genkit tool.

A retriever will use the embedder interface to embed a Document.
We may want to have a way for the retriever to invoke an embedder action,
so that tracing works nicely.