[Go] implement streaming (#108)
Any Action can stream its results back incrementally instead of
returning them all at the end. The same is true for Flows.

Modify Action and Flow to take a callback to be used for streaming.

The dev server runAction route supports a boolean "stream" query
parameter, which if true sets the streaming callback to write
lines of JSON to the ResponseWriter. This matches the js implementation.

For now, assume embedders, retrievers and indexers do not stream.
Generators can stream, and they should, but this PR just lays
the groundwork.