[Go] move most of genkit package into core
This PR greatly simplifies the genkit package, limiting it to
the symbols that Genkit app developers, as exemplified by
the programs in the "samples" directory, would need.

To accomplish this, it moves most of the code to a new package, named
core. The core package, rather than the genkit package, is now imported
by the ai package and plugins. End-user applications should not normally
require it.

This overlapped with #229, so unfortunately those (minor) changes
are incorporated here as well, in a slightly different form.