[Go] fix go generate
Some new types were added, and a new aspect of JSON Schema
that the code generator didn't know about.

The new aspect is that a "$ref" can refer to a property
inside a schema. The code generator still doesn't understand
these, but the config file handles them by ignoring them or
replacing them with another type.