[Go] support sending images to googleai/vertexai
The googleai and vertexai API support gs URIs or raw image bytes.
The genkit flow doesn't permit raw bytes, so use base64 encoding
of data URIs as the TypeScript code does.

This revealed that the JSON validation code didn't support Part
correctly, as the JSON marshaling didn't match the type as seen
by the JSON reflection code. Fortunately the JSON reflection package
makes it easy to specify a different schema to use for any type.

Somehow menu.jpeg was submitted as an empty file, so fix that.