[Go] cross-language test: googleai embedder
Add a test designed to be run in different languages.

This uses a cross-language test framework I'm developing,
github.com/jba/xltest.

At its heart is the file testdata/googleai-embedder.yaml, which defines
the test in a language-neutral way. The test refers to a json file
with the expected trace.

The code in tests/googleai_test.go reads the test file and runs
the test. The main challenge is making two traces comparable,
which requires canonicalizing their span IDs.

To make sure we are testing the complete logic of an embedder action,
we pass a mock HTTP client to the googleai plugin, so we control
the interaction with the service.

Capturing traces cleanly also required some changes to the tracing
system, to allow installing a trace store temporarily and then
removing it.