[Go] make ActionType internal (#360)
ActionType is really an implementation detail. By making it internal,
not only do we reduce clutter in the user-facing API, but we also
make it impossible for 3P code to register any actions of a type
that they shouldn't. For instance, now the only way to register an
embedder is with ai.DefineEmbedder.

The exception is custom actions, so we provide DefineCustomAction.