[Go] suggestion: associate type params with Action Types
Define a type called Assoc that links the Go type of an action
to a value of the ActionType enum. That is, a line like

   const fooAssoc = atype.Assoc[Action[int, bool, string]](atype.Foo)

encodes the relationship between the ActionType enum value Foo and
actions with type parameters [int, bool, string].

The main benefit is that the relationship needs to be written only once
in the program. Previously it appeared twice: in the Define function and
in the Lookup function.