[Go] proposal: revised API
In this API, configuration (defining actions) should still happen
at the beginning of the program, but can be done incrementally by
calling functions on a plugin. For example:

    err := googleai.Init(...)   // creates state common to all actions
    m1 := googleai.DefineModel("gemini-1.0-pro")
    e1 := googleai.DefineEmbedder("emb-01")

When everything has been configured, the program must call genkit.Init.
This has two effects:
- Attempts to register additional actions will fail.
- Servers are started.

This API combines the convenience and clarity of building and storing
actions individually with the requirement that everything that the
program needs is registered early.