[Go] proposal: AI types have methods
Calling Retriever on a retriever as a method is more natural
in Go than calling a top-level function, as we currently have.

This PR changes Indexer and Retriever from being aliases of Actions
to being their own types, allowing methods.

If we like this, then a followup PR would do the same for
Model, Embedder and possibly Tool.