[Go]  Indexer and Retriever types
Calling Retrieve on a retriever as a method is more natural
in Go than calling a top-level function, as we currently have.

This PR changes Indexer and Retriever from being aliases of Actions
to being their own types, allowing methods.

Subsequent PRs will do the same for Model, Embedder and possibly Tool.