[Go] improve core action API
Define a single, general newAction function.

Define a single fully general method for defining an action
in a registry. All other Define functions call it.
Export it so the genkit package can use it to define flows.
It's unfortunate that we have to export it, but code outside the
module can't call it anyway because it takes a registry, which
has an internal type.