[Go] improve API (#523)
This PR consists solely of refactoring. There are no behavior changes.

Address the following warts on the API:

- Although flows are a key concept in Genkit, and the genkit package
  includes a DefineFlow function, the Flow type itself lived in the core
  package.

- The core package contained several exported methods that were intended
  only to be called by the genkit package, a clear sign that something
  is off in the API design.

The actual visible changes to achieve this are small, but a lot of
code was moved. Most of the changes involve moving unexported symbols
into internal packages so they could be used by other packages.

The main API changes are:

- Flow is now in the genkit package.

- There are no more InternalXXX symbols in core.

There are some other minor changes, like the removal of the NoStream
type.