[Go] test picoschema with yaml test cases
Use the yaml file of test cases to test the Go picoschema implementation.

This uncovered a few problems. Some new features were implemented in
JS but not in Go; they are implemented here.

More seriously, it was discovered that the JSON schema package we
are using, github.com/invopop/jsonschema, does not support
arrays for the "type" field, and probably never will (see
https://github.com/invopop/jsonschema/issues/134). Tests that
require that are skipped.