[Go] factor out Gemini plugin common code
The two Gemini plugins share a lot of code, but we can't write
a common package because the genai packages they import are different.
Instead, we copy code from one package to the other.

The copy command in this PR is a simple tool that copies sections
of one file to another.

We mark up the googleai and vertexai files to use this tool, treating
the googleai files as the source of truth.

There is more code to factor out, and we will do that in later
PRs. This is just a start.