[Go] replace IsKnownModel with IsDefinedModel (#615)
IsDefinedModel is more general. You can call it right after Init
to determine if a model is known, or any time to see if you've already
defined a model, and avoid the panic from DefineModel.

Also remove KnownModels. It's unclear if it's useful. If it is, we can
always add it later.

Do the same for embedders.