BREAKING CHANGE!(js): Rip out onFlow and firebaseAuth. Update all docs (#1798)
onFlow is being removed in favor of onCallGenkit in firebase-functions/https. This allows users to use genkit consistently across all environments. Web frameworks (e.g. express, next, firebase-functions) use simple handler functions to use a flow that is defined abstract of the transport protocol.

firebaseAuth is being temporarily removed because it is tightly coupled with onFlow. It will be added in the future as generic middleware that can be used by @genkit-ai/express and @genkit-ai/next to both allow easy enforcement of auth/appCheck and to provide a contex to the flow that is consistent with the one provided by onCallGenkit.

Co-authored-by: Daniel Lee <taeold@gmail.com>
Co-authored-by: Michael Doyle <michaeldoyle@google.com>