feat: add a genkit cli convenience wrapper script to run the repo-local version of genkit cli #1850
RATIONALE:

On some systems, the `npm link` step in `pnpm run setup` fails to
execute owing to a lack of permissions and elevated privileges
are not available. Therefore, this wrapper script is a convenience
feature to allow using the repo-local version of the genkit cli
without having to symlink it into system directories.

ISSUE: https://github.com/firebase/genkit/issues/1850

CHANGELOG:
- [ ] Add a convenience wrapper script that does not require `npm
  link`ing the genkit CLI thereby avoiding permission problems
  faced on certain operating systems.