feat(py/genkit): add a supplementary implementation of GenkitBase to try out a few scenarios
NOTE:

We do not switch the implementation in this PR. It is a WIP.

RATIONALE:

After having played around with several different strategies to manage
all the following 4 scenarios:

| Lifespan  | Environment |
|-----------|-------------|
| Short     | Dev         |
| Short     | Prod        |
| Long      | Dev         |
| Long      | Prod        |

I've come to understand that the cleanest way to handle this is to have
a `run` method on the `Genkit` instance and avoid the use of threads to
manage multiple loops because that can cause weird errors about
scheduling futures after the loop has closed since one would need to be
aware of the loop we're running on.

Removing all threading simplifies the implementation of the Async Genkit
instance to be single-threaded and using `anyio.TaskGroup`s to manage
managing multiple coroutines while blocking at the most appropriate
point to allow users to debug code using the dev server.

The use of `anyio.TaskGroups` is to be compatible with Python 3.10.
Python 3.11 introduces its own `asyncio.TaskGroup`.

In a production setting, we just delegate to running the entry-point
coroutine using `uvloop.run` (preferred if available) or `asyncio.run`.
In a developer setting, we initiate a task group to manage multiple
coroutines (a reflection server) and a user-provided coroutine that may
be short/long-lived.

CHANGELOG:
- [ ] Use `uvloop` in production if available.
- [ ] Add a `run` method on the `Genkit` instance
      to start the main entry-point function.