Fix logs tests from google cloud plugin.
There was an 'it.only' in one of the tests. Removing it uncovered that the two describe blocks couldn't actually both run successfully because they were each enabling telemetry when this can only happen once.  I rearranged the tests to make sure that we only initialize telemetry once, but this uncovered another problem, where only one of them was enabling JSON format for the logs, so I updated everything to use structured JSON format.